Separated React and plain JS code in auth.js, and moved auth components to their own subdirectory
