+++ frontend/components/CardItem.jsx
@@ -15,7 +15,7 @@ import AddMyPlantDialog from './AddMyPlantDialog';
 import { SERVER_ADDR } from '../server';
 import PlantImage from './PlantImage';
 
-const { authFetch } = require('../auth');
+const { authFetch } = require('../api/auth');
 
 class CardItem extends React.Component {
   constructor(props) {
